Rename module
- node_routing -> network
