--- app/main/node_routing/__init__.py
@@ -1,3 +1,0 @@
-from .nodes_network import NetworkManager
-
-network_manager = NetworkManager()
